msm: basic framebuffer support
This framebuffer supports display devices on the primary or secondary
MDDI controller interface, uses the MDP to handle DMA to the displays,
and lays the groundwork for the MSM framebuffer architecture.

The git Author: doesn't well reflect that this one was definitely a group
effort.  Qualcomm provided much of the reference code (in particular the
various init / scaling tables), Brian did the basic driver framework, and
Arve did a lot of refinement and cleanup and added vsync support.

Signed-off-by: Brian Swetland <swetland@google.com>