usb_function: msm_hsusb: watchdog to detect PHY failure
The PHY can (and sadly does) get into a non-responsive state.
Hopefully there will be a better way to detect and recovery,
but to start with we spin up some delayed work to try to read
the PHY interrupt status register every 5 seconds.  In the
event of a read timeout (indicating the PHY is wedged), we
force a USB reset (which will reset the PHY provided the
appropriate hook is provided in platform_data).

Also provide an interlock (the running flag) to prevent
two resets from stepping on each other and a tally of the
number of PHY failure & resets that have occurred in the
"status" debugfs file.

Signed-off-by: Brian Swetland <swetland@google.com>