pmem: adding support to modify backing of suballocated regions, cached pmem
Suballocated regions backing pages can now be changed through a second
call to suballocate.  This change will be transparent to the client, the
same fd and vm addresses will now point to new memory.  To support this
clients may now mmap beyond the size of their allocation, all of these
addresses will map to the same garbage page.

This change also adds support for cached pmem.  Pmem backed userspace
maps will be cached unless the driver is opened with the O_SYNC flag.
The MDP will automatically flush and invalidate the region in question
when a hw accelerated blit is requested

Signed-off-by: Brian Swetland <swetland@google.com>