pmem: Working pmem for 2d acceleration with caching and mdp
This version of pmem allows the first (master) mapper to map
the entire memory space.  Reopening the device will get you a
(client) mapping that points to all garbage pages.  Connect
the two mappings by calling the CONNECT ioctl on the client
cd passing the master fd as an argument. Pages in the master
can be made visible or invisible in the client by calling the
 MAP and UMAP ioctls on the client fd.

Signed-off-by: Brian Swetland <swetland@google.com>