pmem: Adding pmem driver to do contigous memory allocations.
This maps 8M starting at 76M for use by the driver.  Users
requiring large allocations should open (/dev/pmem) and mmap.
The mmap will prompt the kernel to allocate space for that file.
The handle for the pmem space is the file descriptor, apps that need
to pass the allocation to the kernel (mdp, adsp, gl) should pass
in the fd.  Drivers can look up the fd using get_pmem_file which takes
the fd and returns the starting and ending physical addresses of the region.

Signed-off-by: Brian Swetland <swetland@google.com>