pmem: Make get/put pmem file safe to call from an exiting process.
The fucntion fget can not be called on the fd table of a process that
is exiting.  This patch includes a version of get/put pmem file that
hold on to the file struct rather than the fd, so another process can
safely release a pmem file at exit.  This is necessary for the adsp
to decrement references to pmem files if it is killed by the low
memory killer.

Signed-off-by: Rebecca Schultz <rschultz@google.com>