pmem: Fixing a bug in release of master fd
When the master fd is released all sub regions are mapped to garbage.
If one of the subfd's was released while the master was closing, the
code would reference a vma that was no longer present.  The code now
holds the mmap_sem during the unmapping preventing the vma's from
being freed while the unmap is taking place.

Signed-off-by: Rebecca Schultz <rschultz@google.com>