msm: clock: adjust delay based on clock speed, stepping improvements
- jiffy values are calculated using the BLACK MAGIC called 'math'
instead of pre-calculated values. I only saw +0.3% error (ie: new
value is higher) in a one or so cases

- baseline clk speed and lpj now initalized, jiffy is set on the
first clock change (to avoid complications due to power collapse on
idle)

- Also I optimized the current stepping @ 384mhz, so now it only
takes 2 jumps instead of 5.  (This will change a bit more once
cpufreq makes it in, but just in case).

Signed-off-by: Mike A. Chan <mikechan@google.com>