msm: smd: use names instead of numbers for smd channel IDs
Take advantage of the SMD channel allocation structure in SMEM
to publish SMD channels by their names.  Clean up the allocation
of the smd_channel struct so that they are allocated during publish,
and are not leaked.

SMD channel lookup and debug code based on a patch from Zach Pfeffer.

Signed-off-by: Brian Swetland <swetland@google.com>
Acked-by: Zach Pfeffer <zpfeffer@qualcomm.com>