msm: clock: Use smaller AHB divider when CPU runs at 256MHz.
When the CPU runs at 256MHz, the AXI bus is set to run at 128MHz.

Using an AHB divider equal to 3 results in the AHB bus running much slower
(256/3 = 85.333 MHz) than the AXI bus.

Reducing the AHB divider to 2 will result in the AHB bus speed matching the
AXI speed. This increases the memory throughtput without a significant
increase in power consumption.

Memory benchmarks when the CPU is running at 256 MHz shows a throughput
improvement of 25-28% for memcpy of 2KB and 4KB. The improvement is 10-12%
for memcpy of 256KB to 4MB.

Acked-by: Bryan Huntsman <bryanh@qualcomm.com>

Signed-off-by: Brian Swetland <swetland@google.com>