[ARM] msm: acpuclock: Remove VDD check during init.
The VDD register reflects the status of the PMIC only after the
application processor has written to it. This is not the case after a
reset or power collapse since it is the modem processor that sets up
the PMIC. Hence, the VDD register is no longer used to verify the
current VDD level during init (boot up).

Acked-by: Steve Muckle <smuckle@qualcomm.com>