usb: function: Fix phy lockup issue with charger detection
Phy gets struck when we access ulpi phy registes in SE1 state
(both D+ and D- are high). As dedicated wall charger will have
D+ and D- pins shorted, accessing ulpi registers when wall charger is
connected is resulting in PHY lockup. To avoid this, remove accessing
ulpi registers in charger detection code.

To detect the usb WallCharger, enable the pullup on D+ and check line state
using link's portsc register. If linestate is 0x11 it is wall charger
otherwise it is connected to hostpc.

Signed-off-by: Vamsi Krishna <vskrishn@qualcomm.com>