ashmem: Don't install fault handler for private mmaps.
Ashmem is used to create named private heaps. If this heap is backed
by a tmpfs file it will allocate two pages for every page touched.
In 2.6.27, the extra page would later be freed, but 2.6.29 does not
scan anonymous pages when running without swap so the memory is not
freed while the file is referenced. This change changes the behavior
of private ashmem mmaps to match /dev/zero instead tmpfs.

Change-Id: I219c487d430e658a647b1dfdbe956021aec0786b
Signed-off-by: Arve Hjønnevåg <arve@android.com>
Signed-off-by: Sireesha <c_schadi@qualcomm.com>