[ARM] msm: timer: reduce rounding error in time sync calculation
Because the frequency of the debug timer is not a multiple of
that of the slow clock, the order of the calculation leads to
a large rounding error.  It, in turn, causes the system time
to jump ahead intermittenly.  The order of calculation is
changed to minimizes the rounding error.

Signed-off-by: Li, Ai <ali@quicinc.com>