serial: msm_serial_hs: Fix race condition in port close
A work is scheduled to start RX from dmov callback function. Before
this work gets executed a port close may happen where RX will be stopped.
Later when the work gets a chance to run, RX is started again which
prevents the port shutdown to happen (RX flush state never becomes
FLUSH_SHUTDOWN).

Hence check for RX state in the work function and do not start RX if it is
already closed.

CRs-Fixed: 172907
Signed-off-by: Pavankumar Kondeti <pkondeti@qualcomm.com>