[ARM] msm: Add PMIC OTG RPC support
2nd instance of the high speed core requires PMIC OTG
transceiver support, since the PHY is on the modem side.

Modem software provides the infrastructure to initialize/
suspend/resume transceiver.

This file contains the support for RPC calls that are need
to initialize/suspend/resume the transceiver as well the
support for handling the call back events.

when the transceiver is in suspend state, event detections
are handled by the transceiver & communicated to the apps
processor through call backs.

Signed-off-by: Chiranjeevi, Velempati <c_cvelem@qualcomm.com>