[ARM] msm: usb: Add phy reset support to 8x50 platform
In order to reset USB PHY, appropriate bits in APPS_RESET register
must be asserted and then de-asserted. PHY RESET functionality is
implemented at Modem side and Apps HSUSB driver uses RPC to request
the modem to reset the PHY. As part of resetting phy, this api will
also turn off the VBUS. This would not work if HSUSB and FSUSB are
to co-exist. This API checks for FSUSB though, but assumes that
FSUSB is enabled on Modem instead Apps, which brings out additional
complexities. Inorder to support both FSUSB and HSUSB together one
good option is to bring phy reset to apps side. As FSUSB is only
supported in 8x50; Add PHY reset functionality to 8x50 platform.

Note: There is a race condition on the APPS_RESET register since the
modem may modify this register as well.  A fix for this race condition
(only potentially encountered during cable connect or disconnect) will
be added at a later date.

Signed-off-by: Vamsi Krishna <vskrishn@qualcomm.com>