diag: Edit initialization of mempool for DIAG clients on Apps processor
DIAG uses a memory pool for clients on Apps processor for sending their
traffic to USB. This pool was initialized when there were no clients.
This has been now changed to initialize only when there are no clients
as well as no outstanding transactions with the USB. There is no point
re-initializing the mempool just if there are no clients, as there might
be some transactions pending with the USB.

Since now the initialization function is called only when count is zero,
there is no need to initialize it to zero, inside the function.

CRs-fixed: 187391

Signed-off-by: Shalabh Jain <shalabhj@quicinc.com>