Remove check for ATAG_CORE_SIZE
Currently, __vet_atags will erase R2 (atags_pointer) if ATAG_CORE_SIZE
is not exactly 5. This check is too strict. It is possible for
other tags to be present even if ATAG_CORE is emtpy. We should only
check to see if ATAG_CORE is present, to allow the size to grow in the
future.

Signed-off-by: David Brown <davidb@quicinc.com>