driver: fix msm_rmnet driver problem with dropping packets
In the event with rmnet_xmit can not write to smd channel, it schedules
a timer based callback only once to attempt the write later. If the smd
channel is clogged when the callback is invoked the timer is rescheduled
once again. This single reschedule of the timer continues until the
write succeeds, essentially introducing uplink flow control in the driver.

CRs-Fixed: 180014
Acked-by: Jimi Shah <jimis@qualcomm.com>
Signed-off-by: Abhijeet Dharmapurikar <adharmap@quicinc.com>