[ARM] video: msm: mdp_lcdc: Do not force EBI1 to max whenever the screen is on
Previously, when the frambuffer was located in the EBI, there would be visual
artifacts if the EBI1/AXI frequency was not forced to 128mhz. This should no
longer be required, and we should let the ARM9 scale EBI/AXI appropriately
during idle or static screen refreshes. The MDP blits will force it higher
when necessary.

Change-Id: Iace4a056abbe50dc7d1ddb70fe3636873e5cfaf3
Signed-off-by: Dima Zavin <dima@android.com>