[ARM] video: msm: kgsl: disable the GPU during periods of inactivity
The GPU is only enabled when there are pending commands in the ringbuffer.
Whenever the last user of the driver leaves the kernel, a timer is started
which will check the emptyness of the ringbuffer and whether or not the GPU
itself is idle. If the conditions are not met, the timer is restarted. We try
aggressively turn off the GPU after slightly more than a frame of inactivity.

Change-Id: I9a9be81d52431ef2defdc22a3df3af2b2b1ce817
Signed-off-by: Dima Zavin <dima@android.com>