[ARM] msm: qdsp6: Invalidate video encoder output buffers before encode
When an output buffer is passed to the DSP, we have to invalidate it first
to make sure no dirty cache lines are later written to the physical memory
while the DSP is using it.

Also fixes problems with cleaning the input buffer cache before the DSP.
Userspace is not setting the size of the input buffer on each ioctl, so
we have to pre-compute the size based on the encoder paramters.

Signed-off-by: Colin Cross <ccross@android.com>
Change-Id: I7a858f0d833e82b296ac7e32eeaff820d52acba9