usb: msm_function: Allow idle/susp pc when usb Wall-chgr is connected
USB driver prevents idle and suspend power collapse(pc)
while usb cable is connected. But when dedicated charger is
connected, driver can vote for idle and suspend pc. In order
to allow pc, driver has to initiate low power mode which it
cannot do as phy cannot be accessed when dedicated charger
is connected due to phy lockup issues. Just to allow idle &
suspend pc when dedicated charger is connected, release the
wakelock, set driver latency to default and act as if we are
in low power mode so that, driver will re-acquire wakelocks
for any sub-sequent usb interrupts.

CRs-fixed: 214122

Signed-off-by: Vamsi Krishna <vskrishn@qualcomm.com>