msm: Cache routines with and without barriers for DMA operations
DMA API providing cache routines with and without barriers.

The current ARM tree does not meet the DMA API needs of Scorpion.
Historical evidence suggests that some APIs (such as a hard invalidate
that will not clean) will never be implemented upstream. In addition,
it is not always clear upon driver code review that memory type mismatches
should force a change in which upstream API the driver should be calling,
leading to coherency-related problems that are difficult to diagnose and
often surface only on rare occasions. This API ensures that the driver
code is clearly, easily reviewable and is always using both the most
efficient and safest method for cache management. An added advantage
is that this API can be used in the future as an abstraction layer: as
the upstream cache management support improves, these functions can be
remapped via macros to their appropriate upstream APIs, while still
allowing the level of control Scorpion requires for maximum driver
performance (such as graphics) in the short term.

Signed-off-by: Suren Eda Naarayana Kulothungan <sedanaar@quicinc.com>