msm_i2c: I2C controller should be in receive-data state during reads.
Between writing address and polling for 'receive-data' state, I2C
controller seems to transition from 'address' state to 'tx-data'
state and then to 'receive-data' state. So polling during reads
for just 'data' state may be incorrectly over when the controller
is in 'tx-data' state and not really 'receive-data' state.
This commit fixes that problem by making sure that I2C controller
is in 'receive data' state and not just 'data state'.

Signed-off-by: Sagar Dharia <sdharia@quicinc.com>