pmem: Fix Bitmap allocator to use correct meta data sizes
The PMEM bitmap allocator makes use of bitm_alloc to keep track of memory
allocation requests. When the pmem device memory usage increases, the
number of bitm_allocs should be increased accordingly. The reallocation
should therefore be in multiples of bitm_alloc sizes. Add a sanity check
to make sure the number of bitm_allocs is within the maximum and also,
intialize the quanta to zero.

CRs-fixed: 216026
Signed-off-by: Naveen Ramaraj <nramaraj@quicinc.com>