diag: Modified USB callback functions for read/write.
The function signature are changed to pass back a single structure
instead of 2-3 arguments. More number of arguments need to be added to
have a better buffer management scheme for DIAG. Instead of changing
function signature each time a new parameter is added, now just the
structure definition will be changed.

Conflicts:

	drivers/usb/gadget/f_diag.c

Signed-off-by: Amar Singhal <c_amars@quicinc.com>