Bluetooth: Hack: Do not use power_save feature.
power_save is a feature to allow HID devices to control the sniff mode.
Unfortunately it also prevents us exiting sniff mode on some A2DP devices
that do not explicitly exit sniff mode themselves, resulting in skipping
audio.

Marcel is trying to finalize a setsockopt() API to control whether power_save
is enabled on a per socket basis. In the mean-time, turn off power_save for
Android, since we do not have official HID support, and this is causing
problems for A2DP which we do support.

(Note the power_save logic is reversed, 0 is on, 1 is off).

Change-Id: I60a3ffb0d2feb56e854dbc4f2e9f4c580489933d
Signed-off-by: Nick Pelly <npelly@google.com>
CRs-fixed: 254730
Signed-off-by: Asutosh Pathak <asutos@codeaurora.org>