ARM: tegra: harmony: add sdhci devices
Add platform devices for SDIO{1,2,4} on harmony, which goes to wifi and
the two card connectors on the board. SDIO4 (the non-bayonet connector)
is wired for 8-bit cards, SDIO2 is not.

Originally from:
Signed-off-by: Colin Cross <ccross@android.com>

SDIO1 pieces from:
Signed-off-by: Gary King <gking@nvidia.com>

Cleanup, move of GPIO setup to sdhci-plftm driver, etc:
Signed-off-by: Olof Johansson <olof@lixom.net>
---

Depends on the driver that has been merged into mmc-next.

Change-Id: Ibe297f9d3abf8cecbd79bd12f1d56ff4a26016fc