msm: smd: Handle tty errors more gracefully
The function tty_prepare_flip_string can return 0 if there was an
allocation failure in the kernel.  In this situation, we want to
break out of the workqueue thread to give the system some time to
piece together some larger blocks of memory, otherwise we will
spin indefinitely in a loop, effectively livelocking the system.

Change-Id: If77cefc013512d6bc2bc1dec3506941ae2e06194
CRs-Fixed: 231873
Signed-off-by: Michael Bohan <mbohan@quicinc.com>