input: pmic8058_keypad: Fix row hold time calculation
Fix the row hold time calculation so that we can get
exact no. to program the related bits in keypad scan
register. Earlier equation was calculating one value less
than the value needs to be programmed in the register.

It also adds minimum and maximum values checks for row hold
time during probe.

Change-Id: Id35bf56e558228103709a375b6c0ac096dbf7e00
Signed-off-by: Trilok Soni <tsoni@qualcomm.com>