defconfig: 7x30: Enable MMC_UNSAFE_RESUME config option
If MMC_UNSAFE_RESUME is not defined, we enable the mmc bus suspend and
resume functions where we deselect the cards on suspend and re-detect
them on resume. With this config option enabled card re-initializations
can be prevented on every suspend resume cycle. However, it is assumed
that the card will not be physically removed and replaced with another
card while the system is in suspend state.

Change-Id: I1ea16624a4802bb2f81bcdb925d219aedbc75c1b
Signed-off-by: Sahitya Tummala <stummala@qualcomm.com>