msm: acpuclock-7x30: Correct PMIC data used for voltage scaling.
The buck regulator "enable" and "pulldown enable" bits of the PMIC
data should not be considered part of the "VREF_SEL" field. Fix this
and update the PMIC data mask accordingly.  Also, update the voltage
formula to support both 0.625V and 0.3125V references.

This fixes detecting successful voltage changes as failures, which
prevented ACPU frequency increases.

Change-Id: I331aa57e1371bfef1ffd7c2b4af35ef4f0329c48
Signed-off-by: Matt Wagantall <mattw@quicinc.com>