8x50: audio: Fix oops on PCM playback end.
When a zero-length PCM file is played via ALSA while a real PCM
file is also being played, the kernel will panic when the stream
ends because of a null buffer.

This change sets the enabled flag to false on close and checks for
a null buffer before dereferencing it.

Change-Id: I46ad7078e086238b7efc84c4ce6e8d8972d54e76
CRs-fixed: 234588
Signed-off-by: Peter Lohmann <plohmann@quicinc.com>