msm: msm_sdcc: Do not check for DATA_BLK_END bit to exit the SDCC ISR.
In the SDCC ISR, when checking for the bits set in MCI_STATUS register
look for only those bits that correspond to interrupts enabled through
MCI_INT_MASK0 register. Interrupts on DATA_BLK_END are not enabled and
this bit should not prevent the control to exit from the SDCC ISR. If
this is not ensured, the control gets stuck in the ISR for longer time
till the DATA_END event occurs, until when, the DATA_BLK_END bit keeps
getting set for every data block that is transmitted.

Change-Id: I6ee01e9f99d41e28b52240db9859b29c31f9e82b
Signed-off-by: Murali Palnati <palnatim@qualcomm.com>
Signed-off-by: Aparna Mallavarapu <aparnam@qualcomm.com>