msm_fb: Added hardware workaround for MDP-3.1
The source image which is drawn on the
famebuffer memory before sending to a display panel
is broken into 2 or 4 smaller regions
if its width and height on the frame buffer
(destination) meets any of the following conditions:

1. Source image splits in case it is drawn on
framebuffer memory with destination width mod 32 is
14,22,30 or 6 and destination format is 32 bpp.

2. Source image splits in case destination height
or width mod 32 is 3, 2 or 1.

A special case is added in above cases for destination width
or height of 3 to avoid breaking of source window to width
and height of 0 respectively.

Scaling to width or height of 1 can lead to hang in the
MDP3.x. Modifies blit IOCTL to allow blitting of images
on framebuffer at destination width or height of 1 only
when no scaling of image is requested by the user.

Scaling of source image to destination size is changed to use
MN scaling option from hardware in case source size is more than
4 * destination size - 4.

MDP_CGC_EN register is set before PPP kick of and reset after
PPP complete. This sets the gating of PPP block and avoid
premature PPP interrupt for very thin ROI blits.

Change-Id: I6e5ad6cb74e294462ceeacd618cd1cf20bbbc7ec
CRs-fixed: 234329, 232908
Signed-off-by: Neeraj Kumar <c_neekum@quicinc.com>