diag: Add circular buffer to enhance debug capability
DIAG routes data to PC or modem processor through USB & SMD channels.
Sometime data gets lost during interactions with these 2 modules. In
such cases, it is important to know which transfer was left incomplete.
Now, each time an operation with these modules is completed, an entry
will be made in this new circular buffer. The last entry in the buffer
will tell which operation was performed last and will help debugging.

Change-Id: I97d4ae47aebaf728707be35f2ac62675df85895b
CRs-fixed: 236028
Signed-off-by: Shalabh Jain <shalabhj@quicinc.com>