diag: Increase buffer size for increase in incoming item size
DIAG receives data from SMD which is subsequently written to USB.
It receives data in a static buffer of size 8K. With changes in
chaining of DSM items on modem side, the incoming packet size can
be much higher. So, in case a bigger size item comes in, the
current buffer is freed and a new buffer for the required size is
created.

Change-Id: If1f1d38501579422bd0c3111e0a9c6ea464810af
CRs-fixed: 236028
Signed-off-by: Shalabh Jain <shalabhj@quicinc.com>