msm: notify server status to RPC client on modem
When modem does sync operation on eMMC card, and if the RPC
server is not availble on Apps, then the modem will crash trying
to send RPC request. At instances such as system shutdown and
failure in server initialization the RPC server will not be
available. At such instances, notify the server status to the
RPC client on modem through a flag in SMEM.

Change-Id: Ie98f36681ea6761f349de870ee839d5fc669c260
Signed-off-by: Sahitya Tummala <stummala@qualcomm.com>