msm: spm: migrate code that sets 7x30 Scorpion voltages to SPM driver
Migrate code that sets the Scorpion voltages on 7x30 from acpuclk to
the SPM driver.  It serves two purposes: 1) it allows the SPM driver
to restore Scorpion voltage to the correct level after coming out of
power collpase; 2) it consolidates code that deals with the SPM AVS
Wrapper hardware block to the same driver.

Change-Id: I6c3f0863926693b734e603ef482f1f0149448f99
Signed-off-by: Ai Li <ali@quicinc.com>