mfd: marimba-tsadc: Make TSSC as sub-device of TSADC
TSSC (Touchscreen controller) resides on MSM and it is memory mapped,
whereas TSADC (Touchscreen ADC) resides on external Marimba chip,
and driven over SSBI bus.

TSADC driver is designed as sub-device driver of Marimba and TSSC
driver is as standard platform bus driver, but there is no linkage
between ADC and TSSC driver because we can't plug the TSSC directly as
sub-device of the TSADC because it is on MSM SoC and memory mapped.

TSSC driver should have parent device pointer pointing to the TSADC's
device so that kernel PM takes care of suspend/resume sequence and calls
the suspend routine of TSADC only after the TSSC has been suspended
and reverse for resume cycle.

We can't use the MFD apis here because MFD apis mandates that memory mapped
resource needs to have the parent memory resource, whereas in our case
TSADC doesn't have memory mapped resource and it is over serial bus, so
we are adding custom subdev API in the driver which is specific for TSSC.

Change-Id: Ic217c8fc22b1acbd71169872d2c0e2b13cbb754d
CRs-Fixed: 236365
Signed-off-by: Trilok Soni <tsoni@qualcomm.com>