msm: adsp: qdsp5v2: Update adsp driver for wakeup in idle power collapse
Problem:
ADSP driver uses IRQ 20(INT_ADSP_A11) to get messages from ADSP. In
the Idle Power-Collapse case the Modem is not waking up Apps because
it is defined as SMSM_FAKE_IRQ. As a result, the Apps processor is not
woken up when the DSP has messages to deliver to it.

Solution:
The Modem detects IRQ 20 as a wakeup source, but INT_ADSP_A11 is
defined as a fake interrupt. Therefore, swap the definitions of
INT_ADSP_A11 and INT_ADSP_A9_A11.

Change-Id: Ic1abaa065a94f6df2d919c73eca09bc989bd8798
Signed-off-by: Suresh Vankadara <svankada@qualcomm.com>