msm: alsa: 7x30: Fix the wrong check in trigger routine.
Problem
The driver checks wrong flag in trigger routine. This results in
wrong state if the interrupt comes from dsp before the trigger start.
The issue is reproducible if we play audio via alsa utils over serial
console.

Impact
The driver does not play audio if there is lot of delay in user space
to dump data from user space.

Solution
Correct check in driver shall resolve the issue as it wouldn't check the
outneed flag if the playback is in non mmap interface.

Change-Id: I39d1891c29e8b4c595e79c474f8170773aebe1d1
Signed-off-by: Asish Bhattacharya <asishb@qualcomm.com>