msm: 7x27: Flush AXI Bus buffer in msm_arch_idle()
On 7x27 the AXI bus buffer should also be flushed whenever a
cache maintenance operation is being performed. This flush is done
by writing to a strongly ordered memory location. Add this operation
to msm_arch_idle() and also take care of preserving the state of
exception flags and the link register.

CRs-fixed: 238082

Change-Id: I6ffdf68752da96ec0b4a8c8957501a667315e402
Signed-off-by: Naveen Ramaraj <nramaraj@codeaurora.org>
Signed-off-by: Sireesha <c_schadi@qualcomm.com>