msm: 7x27: Flush AXI Bus buffer in msm_arch_idle()
On 7x27 the AXI bus buffer should also be flushed whenever a
cache maintenance operation is being performed. This flush is done
by writing to a strongly ordered memory location. Add this operation
to msm_arch_idle() and also take care of preserving the state of
exception flags and the link register.

Change-Id: I787ccaee2516a1d987d0317d3840a6fead1b94cd
CRs-fixed: 238082
Signed-off-by: Naveen Ramaraj <nramaraj@codeaurora.org>
Signed-off-by: Praveena <c_ppachi@qualcomm.com>