msm: adsp: qdsp5v2: Increase retry duration to 5msecs.
When DSP is not ready to accept a command, ADSP driver retries
for 40 times with 50 usecs delay each time which amounts to
2 msec in total. But when DSP is busy processing commands it may
take more than 2 msec to respond to requests from application
processor.

Hence updating the number of trials to 100 with 50 usecs delay each
time thus increasing the maximum delay to 5 msecs.

Change-Id: Ie2f3a76d7bb33a81a47b62d6fcaac87ffcfbbd88
CRs-fixed:242965
Signed-off-by: Suresh Vankadara <svankada@qualcomm.com>