msm: 7x27: Setup Zero Page required for barriers and cacheops
A Zero page is required to simulate a write to a strongly ordered
memory location to flush the AXI bus buffer in 7x27. The various
barriers and cache operations depend on this page being available
when write_to_strongly_ordered_memory() is invoked. Since barriers
can be called from interrupt contexts this zero page should be initialized
prior to any of these calls.

CR's Fixed: 238082

Change-Id: I62d0a5299a352fb1c8c2e4dcd6311e1731a16289
Signed-off-by: Naveen Ramaraj <nramaraj@codeaurora.org>
Signed-off-by: Sireesha <c_schadi@qualcomm.com>