mm: Make memory hotplug aware of memmap holes
This patch prevents memory hotplug from marking pages of the memmap that
only reference holes in the physical address space as private. Some
architectures (including ARM) attempt to free these unneeded parts of the
memmap, and attempting to free a private page will throw bad_page warnings
and tie up the memory indefinitely.

This patch also allows early_pfn_valid to be architecture specific and
defines it for ARM. The definition for ARM takes into account memory banks
and the holes in physical memory.

CRs-Fixed: 247010

Change-Id: I52584518c4349cf0cb6d625d1ee350bf27db8d68
Signed-off-by: jesset@codeaurora.org