[ARM] tegra: clock: Simplify recalculating clock rates
This patch moves the clock rate recalculation into each of
the generic clock operations that can affect the rate, and
makes all of the clocks use the generic mul and div state
variables to store their relation to their parent clock.
It also removes the now-unused pll rate variables and the
per-clock recalculate_rate operation as they are now all the
same.

Change-Id: I134b9158c06bcab4ea58e4275772a1a1310b7a70
Signed-off-by: Colin Cross <ccross@android.com>