[ARM] tegra2_clocks: fix issues in the tegra2 clock code
UARTs do not have dividers in the clock-and-reset block on Tegra;
the dividers are in the UARTs themselves. remove the DIV_U71 flag
from the UART clocks and extend periph_clk_set_rate to handle
rate == parent->rate when !DIV_U16 and !DIV_U71, and increase
the max rate to 216MHz to match PLLP.

increase the max rate for EMC and PLLM to 800MHz (400MHz DDR), and
add a PLLM table for 666MHz (333MHz DDR).

update CPCON values for PLLM table to match spec

Change-Id: I8f74c42fdebd16799e8d6670813afcada4435f2d
Signed-off-by: Gary King <gking@nvidia.com>