CHROMIUM: make dm= boot path wait for devices
Prior to configuring mapped drives, we attempted to pre-resolve devices
out of the table configurations using the same logic as the other init
code.  However, like md, this may occur before devices are available.

This means that dm booting to a USB device will likely fail.  This
change adds root_wait-style code during device name resolution.  If the
device is not present, then the system will be blocked anyway.  If it
resolves after some period, then we can proceed. If the device resolves
immediately, then we never use this code.

It's worth noting that this will NOT wait for major:minor specified
devices.  I think that's probably okay, but it could be added if it makes
sense.  (forking this off to a separate function which rewalks looking for
' /dev/' or  ' digit*:digit* ')

Signed-off-by: Will Drewry <wad@chromium.org>

TEST=built a kernel and booted a verified rootfs off of usb with it
BUG=chromium-os:327

Change-Id: I350b4deceeeff3c499b4c3e00e4d7ae0645f8b16

Review URL: http://codereview.chromium.org/3059004