msm: audio: 8k: Seperate the DSP buffer read from the userspace read call
In the recording case, the DSP buffer read call is sent in the userspace
read call after copying the previous read buffer to the user space.
This is blocking the DSP read calls whenever the system is loaded. Hence,
moving the DSP read calls to a seperate kernel thread which keeps on
sending the next DSP read call when the event for the successful
completion of the previous read is received. The contents returned by
the DSP are copied to an intermeidate buffer here. The contents of these
intermediate buffers are copied to the user space buffers during the
userspace read calls.

Change-Id: I5140ad982e269a4d89fe6558288e6d7bd8382d11
CRs-Fixed: 241221
Signed-off-by: Pradeep Jilagam <pjilagam@codeaurora.org>
Signed-off-by: Sireesha <c_schadi@qualcomm.com>