diag: Dynamically change size for the pkt registration table
DIAG keeps track of processes using its pkt req/rsp services on
apps processor using a  static table. There is a limit as to how many
registrations can be made at a time. This limit is now made flexible
by increasing the size of the table dynamically as and when required.
However, there is a still a threshold beyond which the size cannot be
increased. And some debug info  will be printed when this threshold is
reached.

Change-Id: I0bc1fdb679cc4012f5c1287f45bf99ad1f549a75
Signed-off-by: Shalabh Jain <shalabhj@codeaurora.org>