diag: Move forwarding functions to work-queue
Currently DIAG forwarding logic occurs in USB and SMD interrupts context
This is because we do not have a dedicated task for Linux DIAG. Hence we
need couple of spinlocks to protect the shared resources. This has
caused deadlock issues on multicore targets like 8660. DIAG has its own
work queue on Linux which is used for several work functions. Moving the
forwarding logic to this work queue will help get rid of spinlocks.

Change-Id: I1bf5cb59e0cb4acf12affd26e18737245c90a068
CRs Fixed: 258037
Signed-off-by: Shalabh Jain <shalabhj@codeaurora.org>
Signed-off-by: Ramakrishna Prasad N <crpn@codeaurora.org>