msm: acpuclock: Vote for bootup PLL
We should ensure that the PLL the acpu clock is running on at
bootup can't be turned off by the modem. Do this by voting for
the pll to be on at boot time. This fixes a crash where the
bootloader hasn't voted for the PLL to stay on and we make a
proc_comm to the modem causing the modem to turn off the PLL the
cpu is running on.

Change-Id: I361e7c29e0aea04b5a5bbd69c7af21cf2247d96c
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>