mfd: pmic8058: replace spinlock with mutex
pm8058_isr_thread() gets a spinlock before it accesses any registers
on the i2c bus. This makes the i2c access to happen in an atomic
context.

The i2c_transfer function either sleeps or returns -EAGAIN in an
atomic context, if the bus is busy. So, in case when the bus is busy
the pm8058_isr_thread gets an error and bails out without handling
the interrupt, only to be invoked again.  The thread using the i2c
bus does not get a chance to complete and a deadlock happens.

Fix the driver to use non-atomic context while doing i2c transfers. This
translates to using mutex instead of spinlock.

Extend this fix to the pm8058-gpio.c, making it to call pm8058_read
and write functions in a non-atomic context.

Change-Id: Ic2f1b6d39892e3b79f1bf7d8d75969ccd57596bd
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>