CHROMIUM: USB: Force use of IO watchdog on EHCI 27cc controllers.
The reason for this is a very long and very sad story involving intermittent
firmware load failures for some USB devices. In particular, if we're talking to
a USB device over an EHCI host controller, we'll /occasionally/ drop an
interrupt, which we think happens because of a bug in the controller itself.
This is normally not a problem, since most drivers support multiple in-flight
URBs, and the chances of dropping a full URB queue worth of interrupts are once
are miniscule. However, miniscule isn't zero - about one time in two thousand,
firmware fails to load on some of our USB devices.

This change forces the use of the IO watchdog, which polls for USB traffic; this
fixes the intermittent firmware load failure at the price of some power
consumption. A proper fix will hopefully be forthcoming from Intel.

TEST=Adhoc
Find a machine with a USB device in it (ideally one whose driver does not
support URB queueing and is not very fault-tolerant). Repeatedly (emphasis on
repeatedly) reboot the machine; eventually the USB device will end up in a sad
state for no apparent reason.

BUG=4561

Signed-Off-By: Elly Jones <ellyjones@chromium.org>
Signed-Off-By: Jason Glasgow <jglasgow@chromium.org>

Review URL: http://codereview.chromium.org/2883064