CHROMIUM: add UUID device selection to dm-verity
This changes addes UUID and UUID+1 based device selection to dm-verity.
If a table provides UUID or UUID+1 as the backing device or hash device
then it will use the found device or the best guess at the +1 partition
after that device.

It allows the firmware to provide the UUID of the kernel partition and
automatically select the root partition (+1) for for kernel-a or
kernel-b.

As a bonus, the refactoring of the device retrieval also let me fold in
device waiting.  If a usb device isn't ready on dm="" setup, it will
rescan if dm_verity.dev_wait=1 is set.

Signed-off-by: Will Drewry <wad@chromium.org>

EST=manual build and use. Requires EFI_PARTITION for the GUID booting
BUG=chromium-os:5081

Review URL: http://codereview.chromium.org/3050041

Change-Id: I2d59ae00bc852a889eaf74a1a82101fb6d6db209