mmc: card: resume host before card resume
There could be a chance where system suspend can fail after the
card is suspended but before host is suspended. And, if the host
is already suspended due to auto-suspend functionality, then the
system is not aware of this and it will not try to resume the host
in case of such failures. Thus, first resume the host if it's already
auto-suspended and then resume the card.

CRs-fixed: 249563
Signed-off-by: Sahitya Tummala <stummala@codeaurora.org>

Conflicts:

	drivers/mmc/card/block.c

Change-Id: I42d88ab0f6651553896c58a260b0419389c7c168
Signed-off-by: Vasantha Kumar <vasant@codeaurora.org>