msm: mmc: configure SDIO wakeup IRQ
All SDCC interrupts are level-triggered, not edge-triggered.
Edge triggering can sometimes cause interrupts to be missed,
especially with the Libra card. Pull high the MSM GPIO that
is used as SDIO wakeup IRQ and then configure it as level low
triggered IRQ.

Wakelock is acquired to abort the suspend process in the scenario
where there is an incoming data and the system is not yet suspended
but SDCC is suspended. The wakelock is released only after the host
is resumed completely and the SDIO event is processed.

Change-Id: Ia20cc55e4a7da7cea653e33f38425a5d078ccfd2
CRs-fixed: 247532
Signed-off-by: Sahitya Tummala <stummala@codeaurora.org>