msm: 7x27: Setup Zero Page required for barriers and cacheops
A Zero page is required to simulate a write to a strongly ordered
memory location to flush the AXI bus buffer in 7x27. The various
barriers and cache operations depend on this page being available
when write_to_strongly_ordered_memory() is invoked. Since barriers
can be called from interrupt contexts this zero page should be initialized
prior to any of these calls.

CRs-fixed: 238082
Signed-off-by: Naveen Ramaraj <nramaraj@codeaurora.org>
Conflicts:

	arch/arm/mach-msm/board-msm7x27.c

Change-Id: I4dc70b74fef0999f916f9916f645be342ea3a6f2
Signed-off-by: Atul Raut <araut@codeaurora.org>