arm: tegra: modifying pinmux for seaboard
Several changes to cleanup pinmux for seaboard. After some testing there are
some corrections that need to be made to the pinmux table. Among these are:

1. A previous change to the pinmux table seems to have been incorrect for cdev2
which caused the 3G modem attached to not function properly. Changing this
seems to have the desired effect of the modem working and nothing else
failing.

2. A Change the DDC signal to RSVD2 so that the I2C can multiplex the signals in
sw correctly.

3. Set UAD to IRDA which is the proper signal to correspond to UARTB.

Signed-off-by: Rhyland Klein <rklein@nvidia.com>
Signed-off-by: Olof Johansson <olof@lixom.net>

Review URL: http://codereview.chromium.org/3130036