msm: rpc: Enqueue reply tags in pending list during write failure
RPC reply tags are not enqueued in pending list when the RPC reply
message write fails. But RPC retransmits the message when there is
a failure. This retransmission fails because the reply tag is not
found in the pending list. So enqueue the reply tag in the pending
list during failure.

Change-Id: I9e97e27c5060c57dcf6affc1b22f785ca2a5b178
CRs-Fixed: 252466
Signed-off-by: Karthikeyan Ramasubramanian <kramasub@codeaurora.org>