msm: Secure Channel Manager (SCM) support.
Implement basic support for SCM communication. SCM uses a
physically contiguous memory buffer as the communication
HEADmechanism. A common command header (struct scm_command) and
response command and response buffers (defined by the headers)
are transparent to the SCM.

A user should call scm_call() with a pointer to the data which
should be passed in the command buffer as well as a pointer to a
buffer to copy the response into.

Change-Id: Ie95906d46b718c5828de10c5e310d270dd2a314e
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>
Signed-off-by: Atul Raut <araut@codeaurora.org>