CHROMIUM: dm-verity: enable target specific error behavior
~~
Patch 1/4: Please take a look across all of them if you can.
While I think we could centralize much of the platform specific logic
in something like a chromeos_behavior module, I'm not sure if it is
worth doing just yet.  These patches show some of the added uses we
might have.  (In addition, I can see us wanting to probe for devmode
v recovery mode at kernel-time as well.)
~~

Allow a target to optionally override the global error_behavior for a
target. This will allow one device to cause a reboot or another to just
return EIOs.  Without this, it is impossible to test bad targets on a
machine with aggressive error behavior.

Change-Id: Ia97acd3c03882d3032d14a2cf8dd13bbcf3d0f4d
Signed-off-by: Will Drewry <wad@chromium.org>

Review URL: http://codereview.chromium.org/3111035