msm: Software workaround for Scorpion Uni 45nm (SC45) Bug
On Scorpion Uniprocessor 45nm cores (SC45) executing an ICIMVAU
and a BPIMVA can cause incorrect entries to be brought in the
TLB with the wrong ASID values but correct virtual to physical
mappings. If any subsequent process that is spawned happens to
share that ASID value, it will make use of the incorrect mapping
in the TLB and can thus fail. This patch enables the recommended
software workaround to fix this problem.

This bug is not applicable to any ScorpionMP or Scorpion Uni 65nm
(SC65) cores.

Change-Id: I89f06ee6e6ce757f23729a52dbaa4dd31ba88da5
CRs-fixed: 251171
Signed-off-by: Naveen Ramaraj <nramaraj@codeaurora.org>