mmc: msm_sdcc: Fix race condition between SDIO ISR and system resume
There is a chance that we get SDIO wakeup IRQ while the system resume
is in progress and before we disable that IRQ in resume. In this scenario,
with the current code the wakeup IRQ is getting disabled twice resulting
in unbalanced IRQ wake disable WARN_ON(). This patch prevents this race
condition by protecting the critical section with host lock.

Change-Id: Ibd0baaf1c3454bba92383002ecfa50b5b6c0f973
Signed-off-by: Sahitya Tummala <stummala@codeaurora.org>