usb: msm_gadget: Allow low power mode when Wallcharger is connected
Accessing USB PHY during SE1(typical state when wall charger
connected) state may cause PHY lockup. This bug was rectified
in 45nm(and later) PHY which is being used in 7x30 and 8x60
targets. Hence, low power mode(which involves phy access) can
be allowed when WallCharger is connected for 7630 and later
targets. Perform PHY reset as workaround for other targets
before accessing it to put in LPM.
However it was observed that,
USB becomes unresponsive after number of plug in and plug outs
when low power mode is initiated and VBUS is present. Actual
root cause of the issue is not yet identified. As a workaround
enable pclk during low power mode if vbus is present

Change-Id: I25393503c43735bb55196c3c0368bb14c4bd0de8
CRs-fixed: 250457
CRs-fixed: 249292
CRs-fixed: 236285
Signed-off-by: Vamsi Krishna <vskrishn@codeaurora.org>
Signed-off-by: Manu Gautam <mgautam@codeaurora.org>