keyboard: handle extra keystrokes during suspend in qci_kbd
During suspend the keyboard controller is still active and can
generate an interrupt if a key is pressed. When the MSM resumes
it does not see the interrupt. The keyboard controller is deadlocked
waiting for the key to be read. When resuming, do a read from
the keyboard to clear the pending keystroke.

Change-Id: I395bb596f83212caccae3126774d4eaa3e90d86c
Signed-off-by: Neil Leeder <nleeder@codeaurora.org>