usb: gadget: f_mass_storage: Add platform driver support
This will be used for configuring vendor, product and release from
board file.As part of INQUIRY command, the host expects the vendor,
product and release info from the device. If the device does n't
send this info, then host ends up showing this device as NULL device
in the device manager.Fix this issue, by sending the product, vendor
and release info,so that host shows the exact details of the device.

Change-Id: I3ef6c43b70398606aafa996e9b12dbe936df873c
CRs-fixed: 246290
Signed-off-by: Chiranjeevi Velempati <cvelempa@codeaurora.org>