msm: timer: cache TIMER_MATCH register
The timer hardware causes any read on the TIMER_MATCH_VAL register not
reflecting the last write on the register if the time between the read
and the last write is too short.  This patch caches the last write and
changes the read on TIMER_MATCH_VAL to return the cached value.

Change-Id: I920c482406d56e59d5868a3963e7f1ce52a21b10
Signed-off-by: Ai Li <aili@codeaurora.org>