msm: timer: cache TIMER_MATCH register
The timer hardware causes any read on the TIMER_MATCH_VAL register not
reflecting the last write on the register if the time between the read
and the last write is too short.  This patch caches the last write and
changes the read on TIMER_MATCH_VAL to return the cached value.

Change-Id: Iaa7edcc84b57b52d277c5f0ff2aa9df532294adb
Signed-off-by: Ai Li <aili@codeaurora.org>