camera: switch power supply for FLUID 12M camera and D02.
LVSW1 doesn't have enough driving strength for the Sony 12M sensor,
its voltage drops below the 1.7V required by the sensor. The solution
is to modify the camera daughter board so that GP16 turns on as
digital voltage for main camera and D02, and GP6 turns on secondary
camera. Meantime, GPIO143 is driving high to put secondary camera in
standby.

Change-Id: Icbde0c6aaa9ad110775e50ee2719a19412f963a0
CRs-Fixed: 253514
Signed-off-by: Shuzhen Wang <shuzhenw@codeaurora.org>