mmc: msm_sdcc: Prevent system suspend when SDIO request is pending
There is a chance where we get SDIO interrupt when function driver
suspend is completed and SDCC host suspend is in progress.  Prevent
system suspend in this case by acquiring a wakelock.  Otherwise, the
function driver resume will not be called to complete that request
and this results in timeout error for an application such as ping.

Change-Id: I8c5436608481b5642b6d668b8de5217aea5db555
Signed-off-by: Sahitya Tummala <stummala@codeaurora.org>
Signed-off-by: Sunil Joseph <sunilj@codeaurora.org>