mmc: msm_sdcc: Prevent system suspend when SDIO request is pending
There is a chance where we get SDIO interrupt when function driver
suspend is completed and SDCC host suspend is in progress.  Prevent
system suspend in this case by acquiring a wakelock.  Otherwise, the
function driver resume will not be called to complete that request
and this results in timeout error for an application such as ping.

Conflicts:

	drivers/mmc/host/msm_sdcc.c

Change-Id: If5d8a665c125cbac280c42873e8414a0d3c5e98c
Signed-off-by: Jessica Gonzalez <jgaona@codeaurora.org>