radio: Fix for DP timeout and Tune response failure issue
The suspend and resume routines in radio-tavarua modified to reduce the
processing time to make sure that the DP timeout does not happen.
This fixes the crash due to this timeout during the stability tests.

Unwanted read of FM SoC Tune control registers on every Tune interrupt
is removed to reduce the i2c bus transactions.
This fixes the intermittent tune failures seen during FM tune stress
tests.

Change-Id: I452653942436392a4549736bbb19ad0d589074b1
CRs-fixed: 256708, 252894
Signed-off-by: Rajiv Kumar <rkumar@codeaurora.org>