msm: msm_sdcc: Reset SDCC in case of command timeout or CRC errors
Reset the controller for any errors to data command which does transfer
in non DMA mode. This ensures that data is flushed out from FIFOs.
Otherwise, there is a chance for any PIO interrupt to come after the
request is processed and this leads access to uninitialized scatter
gather list.

Change-Id: I342c55926a74e4d93c66dc7afd2fb827c4a871c4
CRs-fixed: 257154
Signed-off-by: Sujith Reddy Thumma <sthumma@codeaurora.org>