usb: msm_otg: Update charger type while using ACA
Charger type is currently defined as INVALID while OTG device is
operating in Host mode. Though with ACA, charging can happen if the
ID state is ID_A, but, charging APIs ignore charging request if
the charger-type is set to INVALID.
Hence, update the charger-type while OTG device is in ID_A state.
Also during ID_A -> ID_GND transition, set the charger-type to
INVALID and stop charging before supplying the VBUS.

Tested this change on 7x30 and verified that the charger-type is
now correctly passed.

Change-Id: Ia237194ff1c453d1661650fe8a1cc83894484f55
Signed-off-by: Manu Gautam <mgautam@codeaurora.org>
Signed-off-by: Sunil Joseph <sunilj@codeaurora.org>