mmc: msm_sdcc: Do not always enable SDCC wakeup IRQ
Enable SDCC wakeup IRQ only when there is no wakeup GPIO
interrupt and when async IRQ bit 22 is set in MCIMASK0. The
MPM is connected to the DATA1 GPIO and not the controller
interrupts and hence, the MPM will be configured to enable
those wakeup sources when it's requested for SDCC IRQ. This
might result in modem waking up Apps whenever the data line
is low (irrespective of async bit 22 in MCIMASK0). Hence, be
cautious and request only for the required slots.

Change-Id: If12dd2717b022a622ef8dd0de06928410810353a
Signed-off-by: Sahitya Tummala <stummala@codeaurora.org>