CHROMIUM: mm: stop coredump from allocating shared memory
A Chrome bug was adding segments of shared memory until it exhausted
virtual address space, then crashing: the coredump ran Out Of Memory
because it was allocating physical to the virtual memory - which is
specifically avoided for anonymous memory, but not for shared memory.

The interface to that particular bug has been removed for now; but it
seems possible that there might be other such triggers, in which case
we would do better not to run Out Of Memory unnecessarily in this way.

Pass the FOLL_DUMP flag down through handle_mm_fault(), so shmem_fault()
knows that it's safe to avoid the allocation in this case.  Though only
FAULT_FLAG_WRITE was passed down before, an audit shows that all sites
have correctly been testing for just that one bit.

When upstreaming, we'll replace follow_page()'s FOLL_flags by
FAULT_FLAGs throughout (as Linus Torvalds proposed a year ago);
but here #define the old FOLL_flags to keep backport patch small.

BUG=chromium-os:6778,chromium-os:7274,chromium-os:7333
TEST=grow.c attached to 7333 Comment 10

Change-Id: If8024069d32a4d52e2255cd9526fc0db63f76d0c
Signed-off-by: Hugh Dickins <hughd@chromium.org>

Review URL: http://codereview.chromium.org/3940003