msm: Add sdio suspend/resume callbacks to sdio_al
Fix a scenario in which The sdio_al thread is woken up before
the SDCC suspend begins. While sdio_al thread is processing the data,
the system suspend continues with SDCC suspend (SDCC clocks will be
turned off). Thus the sdio_al work is never done and is stuck forever.

The solution is that sdio_al will not allow going into suspend in case
it it awake (meaning it has pending data)

Change-Id: Ibb73c5f563d2ed15a8ae7c5ea50e8193e8f52c2b
Signed-off-by: Maya Erez <merez@codeaurora.org>