Input: serio - clear serio event queue after sysfs driver rebind
When rebinding a serio driver via sysfs 'drvctl' it is possible for
an interrupt to trigger after the disconnect of the existing driver
and before the binding of the new driver.  This will cause the serio
interrupt handler to queue a rescan event which will disconnect the
new driver immediately after it gets attached.

This change clears the serio event queue after processing the drvctl
request but before releasing the serio mutex, which will clear any
queued rescans before they can get processed.

I considered other options before settling on this fix including
disabling interrupts during the rebind with serio->stop/serio->start,
but that needed some tweaks to i8042.c in order to work which I think
made it more risky than this approach.

BUG=chrome-os-partner:1318
TEST=Manual validation with instrumented kernel.

Kernel instrumented by adding a printk to serio interrupt handler
that indicated a rescan event was queued and to the drvctl handler to
indicate that the event queue was cleared.

Reproduction involves issuing rebind of trackpad serio port from
psmouse driver to serio_raw driver while generating trackpad input
to trigger interrupts.  For ChromeOS device:
  echo -n serio_raw > /sys/devices/platform/i8042/serio4/drvctl
Then check dmesg to see if log messages indicate interrupts were
fired to trigger a rescan and that the i8042/serio4/driver is correctly
attached to the serio_raw driver instead of psmouse.

I also verified that this change fixes the trackpad problem on ChromeOS
devices as documented in partner issue 1318.

Change-Id: I54bc4bdd0cecc5394ac5d505a26f5b1267e360c2
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>

Review URL: http://codereview.chromium.org/4170009