mmc: msm_sdcc: Fix race between msmsdcc_irq and set_ios methods
The variable host->clks_on needs protection with host lock anytime
it is accessed. Otherwise, it could lead to a scenario where SDIO
async IRQ occurs as soon as clocks are disabled but before the clks_on
is updated. Thus, the irq handler may not see the updated clks_on value
resulting in race condition.

CRs-fixed: 262766
Change-Id: I3ae5e295d53799c4fc537fed9e9638c7fd324b2b
Signed-off-by: Sahitya Tummala <stummala@codeaurora.org>