CHROMIUM: tps6586x: Remove device ID check.
The PMU part seems to be in flux, and the device ID keeps changing,
though it seems to always be functionally the same.
Temporarily remove the failure on device ID check until this settles,
so that we can have a working PMU.  Once it settles, we can re-enable
this device ID check as a failure, and make sure we have the proper
device IDs to check against.

Change-Id: If43e335345678f080e9b8fc7c936545c06ac40e1

BUG=none
TEST=Made sure PMU driver comes up okay, and regulators get instantiated.  Important for camera work since LDO6, which is where the camera board sits, is turned off by default.
Signed-off-by: Andrew Chew <achew@nvidia.com>
Signed-off-by: Olof Johansson <olofj@chromium.org>

Review URL: http://codereview.chromium.org/4089007