msm: 8650a: Change pmem to ~4MB
pmem was earlier reduced to ~2MB because ashmem
is now used. However, the video player holds
one buffer when the device is put into standby.
While resuming the device it tries to acquire an extra
pmem buffer. This causes gralloc to run out of memory
for certain high resolution clips and thus show a
monochrome screen for a few moments until the buffer is
released. Hence, increasing the pmem size fixes the issue.

Change-Id: I4d7ccdf86964bdc5ea35f991e1a2d55982211e10
CRs-Fixed: 258977
Signed-off-by: Naseer Ahmed <naseer@codeaurora.org>
Signed-off-by: Sunil Joseph <sunilj@codeaurora.org>