USB: msm_otg: Allow LPM with wall charger connected
Accessory charger detection (Id polling) is turned off in LPM.
So LPM is not allowed when B-session valid and Id are true to
detect an accessory charger.  When wall charger is connected
B-session and Id are true. But there is no way of connecting
an accessory charger.  So allow LPM with wall charger connected.
Ensure Id polling timer is turned off before entering into LPM.

Change-Id: I190f36cce91519aba82764198c24c81b67ea67a9
CRs-Fixed: 262688
Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>