msm: rpm-regulator: add API for selectable sleep-set regulator requests
Currently, all regulator requests are sent to the RPM to modify the
active set.  Values for the sleep set are inherited from this set
because they are left uninitialized.  Some drivers, such as acpu,
require that their vote only count for the active set and not the
sleep set.

Add an API that allows for regulator voltage setting in which a sleep
flag is used to indicate if both the active and sleep sets should be
modified.

Change-Id: I8b641c129751d52316844c3ffa628423a98fc86c
Signed-off-by: David Collins <collinsd@codeaurora.org>