USB: msm_otg: Enable PHY comparators in LPM with wall charger
PMIC notifications are designed to receive VBUS high event only. We
should rely on PHY comparators for VBUS low interrupt. This does
not matter if we are connected to host. Because bus suspend is
not implemented. But if we don't enable PHY comparators and allow
LPM when wall charger is connected, we will not detect charger
disconnection.

Change-Id: I896547bce747525ad46d0a7d36fd903a1a494957
Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>
Signed-off-by: Ramakrishna Prasad N <crpn@codeaurora.org>