USB: msm_otg: Enable PHY comparators in LPM with wall charger
PMIC notifications are designed to receive VBUS high event only. We
should rely on PHY comparators for VBUS low interrupt. This does
not matter if we are connected to host. Because bus suspend is
not implemented. But if we don't enable PHY comparators and allow
LPM when wall charger is connected, we will not detect charger
disconnection.

Change-Id: Iad792bb24d42665fa6b0dc55e94214c20d068e27
Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>
Signed-off-by: Sunil Joseph <sunilj@codeaurora.org>