msm: adsp: qdsp5: Increase retry duration to 15msecs.
When DSP is not ready to accept a command, ADSP driver retries
for 100 times with 50 usecs delay each time which amounts to
5 msec in total. But when DSP is busy processing commands it may
take more than 5 msec to respond to requests from application
processor.

Hence updating the number of trials to 300 with 50 usecs delay each
time thus increasing the maximum delay to 15 msecs.

Change-Id: I63d6caa607096401e020c93288334ff96ac32f81
Signed-off-by: Suresh Vankadara <svankada@codeaurora.org>
Signed-off-by: Ramakrishna Prasad N <crpn@codeaurora.org>