msm: board-msm8x60: Disable pull down for regulator 8058_l14
Due to a hardware bug on 8x60 boards, disabling regulator
8058_l14 results in repeated spurious interrupts on the
SDC3_CARD_DET_SW GPIO pin.  This causes the SD card driver
to repeatedly try to initialize a nonexistent SD card.
The driver prints a message into the kmsg log each time
this happens.

Disable the pull down for regulator 8058_l14 so that the
SD card detection line is not pulled down via 8058_l14.
This will allow 8058_l14 to be turned off for power savings
and will eliminate the associated spurious interrupts.

Change-Id: I8fa8e70e847fd71fcd38bdfefbc84b514e5d6d28
Signed-off-by: David Collins <collinsd@codeaurora.org>