msm: board-qrdc: change specific to 8x60 v1 for msm bus scaling driver
Pull in the same change made to 8x60 board into qrdc board.
Add the bus topology details for 8x60 v1. The nodes attached to each
fabric are added in the board file. The fabric and slave clocks,
default tiers and bus widths are also initialized. The bus driver will
use this data to construct the topology.

Change-Id: I4b61d9519120e3405f27e04f77f75647b6262b96
Signed-off-by: Jin Hong <jinh@codeaurora.org>
Signed-off-by: Praveena Pachipulusu <pveena@codeaurora.org>