staging: android: lowmemorykiller: fix infinite loop in the notifier chain
The same notifier could be added in the notifier chain if the
shrink function is called concurrently. This causes the notifier
chain to loop around itself and the code is stuck there.

Use the deathpending pointer to check if it is safe to add the
notifier in the chain.

Change-Id: I60abd2bda5a47fb90ca68604114d154f390b9012
CRs-Fixed: 264788
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>