CHROMIUM: NET: USBNET: move tegra 32 byte alignment issue
We need to ensure the sk buf is 32 byte aligned, and it seems that the
asix controller touches the sk buffer in the tx_fixup callback that
changed this alignment. The misaligned buffer gets stuck in the usb
controller and doesn't get issued, causing the timeout.

BUG=https://code.google.com/p/chrome-os-partner/issues/detail?id=1672
TEST=can browse web steadily without seeing the problem

Change-Id: I2efeaa00766c1611feecccf36c59946b5798a691
Signed-off-by: Rhyland Klein <rklein@nvidia.com>
Signed-off-by: Olof Johansson <olofj@chromium.org>

Review URL: http://codereview.chromium.org/5211005