rtc: alarm: Add a API to update the alarm time information
In cases when the system time is updated directly (without
the knowlege of the alarm driver), the new time information
does not get updated in the alarm data structure. This change
adds a API to update the alarm structure.

Change in the rtc-msm and rpc_server_time_remote driver to
call this API after it receives a call back from the MODEM
to update the system time.

Change-Id: Iac319f94724011713b063b1e769e3af055639afe
CRs-Fixed: 266260
Signed-off-by: Anirudh Ghayal <aghayal@codeaurora.org>
Signed-off-by: Ramakrishna Prasad N <crpn@codeaurora.org>