msm: acpuclock-7x30: Request max ebi1 at max cpu frequency
DDR1 runs slower than DDR2, but we were only requesting the max
rate supported for DDR1. This means DDR2 is running slower than
it should be when apps is running at the highest CPU frequency.
Request UINT_MAX, which will notify the remote side to set the
DDR frequency to the max that is supported for whatever DDR part
is present.

Change-Id: Ibab8267112dee100d6885b567b5c36c32105bb79
CRs-Fixed: 267050
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>