rtc: alarm: Add a API to update the alarm time information
In cases when the system time is updated directly (without
the knowlege of the alarm driver), the new time information
does not get updated in the alarm data structure. This change
adds a API to update the alarm structure.

Change in the rtc-msm and rpc_server_time_remote driver to
call this API after it receives a call back from the MODEM
to update the system time.

Change-Id: I7354046c613067f9e1bb916968e5e28d9e483623
CRs-Fixed: 266260
Signed-off-by: Anirudh Ghayal <aghayal@codeaurora.org>