CHROMIUM: config: Enable Ath6kl driver in ARM config
This is not enabled for x86 for now as I do not have a way of testing
this driver on x86. The driver is enabled by CONFIG_ATH6K_LEGACY and
this has been moved to the architecture-specific config files instead
of the main Chromiumos config file.

BUG=chromiumos-10162
TEST=build kernel and check that ath6kl driver is present.
Boot Seaboard and check the same

Change-Id: I610f827dcce87f76a6516ceb67501298de02f215

Review URL: http://codereview.chromium.org/5719006