Bluetooth: Add flag to track managment controlled adapters
This patch adds a HCI_MGMT flag to track adapters which are under the
control of the management interface. This is needed to make sure that
new kernels will work with old user space versions. I.e. behaviour which
could break old user space versions (but is needed by the management
interface) should not be exhibited when the HCI_MGMT flag is not set.

Signed-off-by: Johan Hedberg <johan.hedberg@nokia.com>
Signed-off-by: Gustavo F. Padovan <padovan@profusion.mobi>
(cherry picked from commit ebc99feba7378349e2bfae7018af062767382f6c)

Change-Id: I27753d516b2f1faa030ca10ac0fed209f37dccd7
Signed-off-by: Mat Martineau <mathewm@codeaurora.org>