tpm: fix panic caused by "tpm: Autodetect itpm devices"
commit 3f0d3d016d89a5efb8b926d4707eb21fa13f3d27 adds a check for
PNP device id to the common tpm_tis_init() function, which in some
cases (force=1) will be called without the device being a member of
a pnp_dev. Oopsing and panics ensue.

Move the test up to before the call to tpm_tis_init(), since it
just modifies a global variable anyway.

Signed-off-by: Olof Johansson <olof@lixom.net>
Cc: Matthew Garrett <mjg@redhat.com>
Cc: stable@kernel.org [2.6.37+]