msm: audio: 8660: Change HPF Bypass Sequence
There is a pop on the TX path when the handset microphone
is enabled due to the HPF being enabled before the
microphone bias is turned on. The pop occurs because when
the micbias is enabled, there is a sudden spike in voltage
coming into the HPF.  The HPF takes up to 200 ms to
discharge after the voltage spike, which causes a pop in
the transmitted audio or voice stream.  By bypassing the HPF
by default and only removing the bypass after the micbias
is on, then there is no pop.

A 15ms delay after the HPF is enabled in the Timpani
profile ensures that all pops are removed.

CRs-fixed: 269726
Signed-off-by: Brad Rubin <brubin@codeaurora.org>
(cherry picked from commit 9cbc32573875b620e1f262329c4662853c264290)

Change-Id: I4f746a1b8ef75e465232ec9b7d1f5e0a7a1f859f
Signed-off-by: Sunil Joseph <sunilj@codeaurora.org>