ARM: vfp: fix fpsid register subarchitecture field mask width
The subarchitecture field in the fpsid register is 7 bits wide.
The topmost bit is used to designate that the subarchitecture
designer is not ARM.

This change has the side effect of causing VFP implementations
other than ARM to be reported as VFPv3 with common VFP subarch
v3. Before this change it would be reported as VFPv1 with
an implementation defined subarchitecture.

CRs-Fixed: 268685
(cherry picked from commit 40b0e5767159cf70af24a762381745a95bb9592d)

Change-Id: I18ecd5bf9069637cb1d8aa2d54d196dbf82213d4
Signed-off-by: Steve Muckle <smuckle@codeaurora.org>