CHROMIUM: config: enable CONFIG_X86_MCE_INTEL and re-generate with kernelconfig.
This was enabled in 2.6.32 but is not in kernel-next.  The Intel-specific
config option opens up the ability to do thermal monitoring in the kernel.
With Mario this is done in the BIOS by routing thermal interrupts to an
SMI but I would like to move this into the OS eventually.

BUG=None
TEST=Boot kernel-next and look for Thermal Monitoring messages.

Change-Id: Iea339b65582fb5447eb463b48baa08de7b2f2afb
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>

Review URL: http://codereview.chromium.org/5863004