msm: qdsp5v2: audio: Change device profiles to suppress noise
Increases audio performance for capless devices in timpani profile
minimizing the level of audible noise,
There is a pop on the TX path when the handset microphone is enabled
due to the HPF being enabled before the microphone bias is turned on.
The pop occurs because when the micbias is enabled, there is a sudden
spike in voltage coming into the HPF.  The HPF takes up to 200 ms to
discharge after the voltage spike, which causes a pop in the
transmitted audio or voice stream.  By bypassing the HPF by default
and only removing the bypass after the micbias is on, then there is
no pop.

CRs-fixed: 270615
CRs-fixed: 270808
Signed-off-by: Santosh Mardi <gsantosh@codeaurora.org>
(cherry picked from commit e7f15bdfd6a7ece9db0051c3325edb2ab2e74ba7)

Change-Id: I35063c48f0d0c52269cd1e8c9c843ce454df3159
Signed-off-by: Sunil Joseph <sunilj@codeaurora.org>