diag: Pass correct parameters to overflow check macro
CHK_OVRFLOW takes as its arguments the start of the
buffer, the start of the location in the buffer where
data is to be copied, the end of the buffer, and the
length of the data to be copied. diag_update_log_masks
is passing a value greater than the maximum size of the
log masks buffer as the third argument to CHK_OVERFLOW,
fix this to be the maximum size. Also update the log mask
size to accommodate bigger mask requests.

CRs-Fixed: 270142
Signed-off-by: Shalabh Jain <shalabhj@codeaurora.org>
(cherry picked from commit be4185cc727340bfdc9d119ccf654894d70d3eab)
Change-Id: Ie302f1c526ef4932a8979e4c48e29f8a76ffedf6
Signed-off-by: Sunil Joseph <sunilj@codeaurora.org>