CHROMIUM: tegra: seaboard: enable 3G modem on ULPI port
- Fix the ULPI transceiver configuration for the USB2 controller :
  * Use the right GPIO (PV1) for reset
  * set the reference clock to 24MHz
- Switch on the 3v3 power supply for the mini PCIe interface :
  * Present the MOSFET as a "fixed regulator"
  * Enable it at startup to be able to enumerate the modem USB interface

Note that the MOSFET is triggered by the PMIC (TPS658621) GPIO3.
This mechanism is not present on Seaboard, but this should be reasonably
as the GPIO is not connected on this board.

A power management policy still needs to be implemented to switch on/off the
modem power supply.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

TEST=Load a firmware on the Gobi modem
BUG=none

Change-Id: I20941ad842b8ca116bb4a4105b8d438b45eb0ba6

Review URL: http://codereview.chromium.org/6305005