Bluetooth: Update the security level when link is encrypted
If the pending security level is greater than the current security
level and the link is now encrypted, we should update the link
security level.

This is only useful for LE links, when the only event generated
when SMP is sucessful in the Encrypt Change event.

Signed-off-by: Vinicius Costa Gomes <vinicius.gomes@openbossa.org>
(cherry picked from commit 03f3ddde93ac540daaf22e98e87680fea018b62f)

Change-Id: Ib13fd61dc623b4b077b8f515736ec561e1f7bbdf
Signed-off-by: Mat Martineau <mathewm@codeaurora.org>