USB: gadget: f_rmnet: Enable endpoints irrespective of smd channel status
Endpoints are being enabled only after the smd channels are opened
successfully. Since we have an interrupt endpoint in this function,
it is not allowing the host drivers to be installed properly,
if the endpoints are not enabled by the time, set config 1 is received.
Hence enable endpoints irrespective of the smd channel open status.

CRs-Fixed: 268645
Signed-off-by: Rajkumar Raghupathy <raghup@codeaurora.org>

Conflicts:

	drivers/usb/gadget/f_rmnet.c

Change-Id: Ifc13e295f59c029d91a8e2f6bf2604020d7b697b
Signed-off-by: Asutosh Pathak <asutos@codeaurora.org>