ASoC: dapm - add route checker for valid DSP paths
Add a DAPM API call to determine whether a DAPM audio path is valid.
i.e. it has a source and sink that are connected via kconrol mixer
and muxes.

This will be used by the DSP core to determine the runtime mappings
between FE and BE DAIs in order to run PCM operations.

Change-Id: I9d570f5e1786808e7ef03dd541585409049ca8e8
Signed-off-by: Liam Girdwood <lrg@slimlogic.co.uk>
Signed-off-by: Patrick Lai <plai@codeaurora.org>