CHROMIUM: power: bq20z75: update driver to current version
This change updates the bq20z75 driver to match the latest
upstream version, with the exception of forcing the returned
current to be positive. That change was done to support a measurement
tool, and therefore not accepted upstream.

The main change is to move the logic to see which properties need to be
returned in ua vs which need to be returned in uw into a more common
place so it can be reused by multiple drivers and be maintained as
properties are changed.

Change-Id: Ia75fa7f5bc467c6b0684fc8c821c557b43dea545

BUG=n/a
TEST=verified driver returned sensible values
Signed-off-by: Rhyland Klein <rklein@nvidia.com>
Signed-off-by: Olof Johansson <olofj@chromium.org>

Review URL: http://codereview.chromium.org/6410096