Revert "CHROMIUM: Import atmel maxTouch driver"
This reverts commit 8a3011a6674d5a3b94f7a260517b45c77945a3fe.

Atmel has submitted new versions of the drivers upstream.  This change is
cleaning up our repo so we can pull the new versions directly from upstream

BUG=chromium-os:11929
TEST=Validated kernel breaks witohut 6488011
  Applied 6488011, verified compile

This change can not go in until CLs
http://codereview.chromium.org/6488011
http://codereview.chromium.org/6489014
are in.

The code will not compile until 6488011 is applied

Review URL: http://codereview.chromium.org/6490012

Change-Id: If4f9b41227aa49d275688d2cbd2923ff97e52a5f