msm: kgsl: fix process_private and memqueue locking
struct kgsl_process_private is shared between device nodes
and it is possible that 2 threads of a single process could
try to access this structure simultaneously. This structure
is now locked.

Additionally, the memory queue for to be freed memory and the
process list are now mutually exclusive, a kgsl_mem_entry can
now be on one list or another, not both at the same time as before.
Memqueue is also moved to struct kgsl_device as it is shared
between 2d and 3d.

(cherry picked from commit c8d400655ac72d07cec1f73f1f5f926e9227fd83)

Change-Id: I158bc5af0d63a64ccba0e344ac7c949aa025e2e0
Signed-off-by: Jeremy Gebben <jgebben@codeaurora.org>