CHROMIUM: video: tegra: Implement tegra_dc_hdmi_set_audio_sample_rate
The SPDIF audio driver needs to inform the HDMI driver of the currently
active audio sample rate, so that the HDMI controller can be correctly
configured for that rate. Add a function to allow this.

There is a single SPDIF controller, which is routed to both display
controllers. Hence, iterate over all HDMI controllers when implementing
this functionality.

BUG=
TEST=
Signed-off-by: Stephen Warren <swarren@nvidia.com>
Signed-off-by: Olof Johansson <olof@lixom.net>

Review URL: http://codereview.chromium.org/6632011

Change-Id: Ife93ac58c1b134ec01461854036322c6e0e848de