CHROMIUM: config: refresh for new drivers and more
Not 100% sure about the broadcom rename, if it should be enabled on
x86. If so we can turn it on later as well.

I disabled NCT1008 everywhere but tegra, since it doesn't build on
arm-generic and x86 platforms don't use it (at this time at least).

Enabled HID_APPLE on ARM, since it's needed for apple mice and I happened to
want to use one.

Change-Id: I6cac8c89803691f89798af7b478a01dd6f906e59
Signed-off-by: Olof Johansson <olof@lixom.net>

BUG=None
TEST=build all configs

Review URL: http://codereview.chromium.org/6623049