CHROMIUM: char/tpm: Add support for Infineon's I2C TIS TPM
[Note: This driver will be reworked by the upstream TPM maintainer
to share as much code as possible with a couple of other i2c tpm's.
While that is taking place, we are checking in this driver to have
something to exercise hardware with. This will be reverted when
the upstream driver is available. -olofj]

This patch adds a new driver for Infineon's new I2C based TPM.
The used protocol is quite similar to the TCG TIS protocol.

Tested on: tegra2 seaboard (arm) , beagleboard (arm)
Known Bugs: Suspend does not work correctly on Seaboard, since I2C bus shuts
itself down too early. Works on Beagleboard

Change-Id: I7e945e2fb4edab4eded6c9cd92dd18a1e947c220

BUG=None
TEST= Running trousers testsuite several times, manual tests
Signed-off-by: Peter Huewe <huewe.external@infineon.com>
Signed-off-by: Olof Johansson <olofj@chromium.org>

Review URL: http://codereview.chromium.org/6530003