usb: msm_gadget: Check both USB state and VBUS during initialization
If an A-device is connected as a part of bootup, host controller
driver configures USB in host mode and enables VBUS to communicate
with the USB peripherals. Due to this BSV bit is set in OTGSC
register. Later as a part of USB gadget composition registration,
DCD state machine wrongly reading BSV bit and if VBUS is valid,
configures USB in peripheral mode. Because of this the target in
host mode failed to work with the external USB peripherals.
So before initializing USB in peripheral mode, check usb state and
if it is in ATTACHED state, configure USB in peripheral
mode otherwise just move the USB peripheral state to OFFLINE.

Change-Id: Ib38b615e705f6427bcef6d6aba809b86f712474a
CRs-Fixed: 278202
Signed-off-by: Rajkumar Raghupathy <raghup@codeaurora.org>
Signed-off-by: Manu Gautam <mgautam@codeaurora.org>