Bluetooth: Add functions to manipulate the link key list for SMP
As the LTK (the new type of key being handled now) has more data
associated with it, we need to store this extra data and retrieve
the keys based on that data.

Methods for searching for a key and for adding a new LTK are
introduced here.

Signed-off-by: Vinicius Costa Gomes <vinicius.gomes@openbossa.org>
(cherry picked from commit 57da7d99e47db5aa513e59e677b1e6e49b495ba7)

Change-Id: Ib39205a119ee94f0341e81c6834c3e015d1baaf2
Signed-off-by: Mat Martineau <mathewm@codeaurora.org>