CHROMIUM: dt: add of_platform_populate() for creating device from the device tree
of_platform_populate() is similar to of_platform_bus_probe() except
that it strictly enforces that all device nodes must have a compatible
property, and it can be used to register devices (not buses) which are
children of the root node.

This patch also modifies MPC5200 support to use the new function.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>