Bluetooth: mgmt: Add local name information to read_info reply
This patch adds the name of the adapter to the reply of the read_info
management command.

The management messages reserve 249 bytes for the name instead of 248
(like in the HCI spec) so that there is always a guarantee that it is
nul-terminated. That way it can safely be passed onto string
manipulation functions.

Signed-off-by: Johan Hedberg <johan.hedberg@nokia.com>
Signed-off-by: Gustavo F. Padovan <padovan@profusion.mobi>
(cherry picked from commit dc4fe30b8675033e538e2dea50be8af9c75f1b6a)

Change-Id: Ifb510eec14565a118541fe19bffac3b2c27d09c1
Signed-off-by: Mat Martineau <mathewm@codeaurora.org>