CHROMIUM: ARM: tegra: seaboard: Set up WM8903 IRQ GPIO before pinmux table
PINGROUP_SPIC contains two pins:
+ PX2, DISABLE_CHRGR (output)
+ PX3, WM8903 codec IRQ (input)

The pinmux module can only configure TRISTATE vs. NORMAL on a
per-group rather than per-pin basis. The group must be NORMAL
since at least one pin is an output. However, we must ensure that
the WM8903 IRQ is never driven, since the WM8903 itself is driving
it, and we don't want multiple drivers. To ensure this, configure
PX3 as a GPIO here, and set is as an input, before the pinmux table
is written, which is when the pins will be un-tristated.

[swarren:
  Added comments, commit description.
  Moved code to seaboard_pinmux_init.]

BUG=
TEST=
Signed-off-by: Rhyland Klein <rklein@nvidia.com>
Signed-off-by: Stephen Warren <swarren@nvidia.com>
Signed-off-by: Olof Johansson <olofj@chromium.org>

Review URL: http://codereview.chromium.org/6697019

Change-Id: Ie63889f441bf209c061605cb57fd300856a3fdb7